+++ inc/woocommerce/integrations.php
@@ -58,6 +58,13 @@ function storefront_woocommerce_integrations_scripts() {
 	if ( class_exists( 'WC_Photography' ) ) {
 		wp_enqueue_style( 'storefront-woocommerce-photography-style', get_template_directory_uri() . '/inc/woocommerce/css/photography.css' );
 	}
+
+	/**
+	 * Product Reviews Pro
+	 */
+	if ( class_exists( 'WC_Product_Reviews_Pro' ) ) {
+		wp_enqueue_style( 'storefront-woocommerce-product-reviews-pro-style', get_template_directory_uri() . '/inc/woocommerce/css/product-reviews-pro.css' );
+	}
 }
 
 /**
@@ -78,16 +85,16 @@ function storefront_woocommerce_integrations_layout() {
  *
  * @since 1.0
  */
-if ( ! function_exists( 'storefront_add_bookings_customizer_css' ) ) {
-	function storefront_add_bookings_customizer_css() {
+if ( ! function_exists( 'storefront_add_integrations_customizer_css' ) ) {
+	function storefront_add_integrations_customizer_css() {
 
 		if ( is_storefront_customizer_enabled() ) {
-			$accent_color 				= storefront_sanitize_hex_color( get_theme_mod( 'storefront_accent_color' ) );
-			$header_background_color 	= storefront_sanitize_hex_color( get_theme_mod( 'storefront_header_background_color' ) );
-			$header_link_color 			= storefront_sanitize_hex_color( get_theme_mod( 'storefront_header_link_color' ) );
-			$header_text_color 			= storefront_sanitize_hex_color( get_theme_mod( 'storefront_header_text_color' ) );
-			$text_color 				= storefront_sanitize_hex_color( get_theme_mod( 'storefront_text_color' ) );
-			$heading_color 				= storefront_sanitize_hex_color( get_theme_mod( 'storefront_heading_color' ) );
+			$accent_color 					= storefront_sanitize_hex_color( get_theme_mod( 'storefront_accent_color', apply_filters( 'storefront_default_accent_color', '#a46497' ) ) );
+			$header_background_color 		= storefront_sanitize_hex_color( get_theme_mod( 'storefront_header_background_color', apply_filters( 'storefront_default_header_background_color', '#2c2d33' ) ) );
+			$header_link_color 				= storefront_sanitize_hex_color( get_theme_mod( 'storefront_header_link_color', apply_filters( 'storefront_default_header_link_color', '#ffffff' ) ) );
+			$header_text_color 				= storefront_sanitize_hex_color( get_theme_mod( 'storefront_header_text_color', apply_filters( 'storefront_default_header_text_color', '#5a6567' ) ) );
+			$text_color 					= storefront_sanitize_hex_color( get_theme_mod( 'storefront_text_color', apply_filters( 'storefront_default_text_color', '#787E87' ) ) );
+			$heading_color 					= storefront_sanitize_hex_color( get_theme_mod( 'storefront_heading_color', apply_filters( 'storefront_default_heading_color', '#484c51' ) ) );
 			?>
 			<!-- column customizer CSS -->
 			<style>
@@ -111,6 +118,21 @@ if ( ! function_exists( 'storefront_add_bookings_customizer_css' ) ) {
 					#wc-bookings-booking-form .wc-bookings-date-picker .ui-datepicker-header {
 						background-color: <?php echo $header_background_color; ?>;
 						color: <?php echo $header_text_color; ?>;
+					}
+				<?php }
+
+				if ( class_exists( 'WC_Product_Reviews_Pro' ) ) { ?>
+					.woocommerce #reviews .product-rating .product-rating-details table td.rating-graph .bar,
+					.woocommerce-page #reviews .product-rating .product-rating-details table td.rating-graph .bar {
+						background-color: <?php echo $text_color; ?>;
+					}
+
+					.woocommerce #reviews #comments ol.commentlist li .contribution-actions a,
+					.woocommerce-page #reviews #comments ol.commentlist li .contribution-actions a,
+					.star-rating-selector:not(:checked) input:checked ~ label.checkbox,
+					.star-rating-selector:not(:checked) label.checkbox:hover ~ label.checkbox,
+					.star-rating-selector:not(:checked) label.checkbox:hover {
+						color: <?php echo $accent_color; ?>;
 					}
 				<?php } ?>
 			</style>