+++ changelog.md
@@ -1,6 +1,7 @@
 # Storefront Changelog
 
 ### *2014.11.25* - 1.1.2
+* **New** - Integration with Product Reviews Pro extension.
 * **Fix** - Layout picker images when using a child theme.
 * **Fix** - Header link color now properly applied to site title.
 * **Fix** - Included styling for `.form-row-wide`.